+++ packages/server/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @mastra/deployer
 
+## 0.10.12
+
+### Patch Changes
+
+- Updated dependencies [b4a9811]
+- Updated dependencies [4d5583d]
+  - @mastra/core@0.10.12
+
+## 0.10.12-alpha.1
+
+### Patch Changes
+
+- Updated dependencies [4d5583d]
+  - @mastra/core@0.10.12-alpha.1
+
+## 0.10.12-alpha.0
+
+### Patch Changes
+
+- Updated dependencies [b4a9811]
+  - @mastra/core@0.10.12-alpha.0
+
 ## 0.10.11
 
 ### Patch Changes